pyproject.toml
[project]
name = "voxel-terrain"
version = "1.0.0"
description = "Voxel-based terrain generation and editing tools for Blender"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.11"                                                   # Blender 5.0 uses Python 3.11+
authors = [{ name = "Your Name", email = "your.email@example.com" }]
keywords = ["blender", "addon", "extension", "voxel", "terrain"]

# Dependencies are only for development - not bundled with the addon
# Blender provides its own Python environment
dependencies = []

[project.optional-dependencies]
dev = [
    # Type stubs for Blender API - use 'latest' for Blender 5.0
    # Check PyPI for available versions: https://pypi.org/project/fake-bpy-module/
    "fake-bpy-module-latest>=20240927",

    # Linting
    "ruff>=0.8.0",

    # Type checking
    "pyright>=1.1.390",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# This is a Blender addon, not a pip-installable package
# We only use pyproject.toml for dev tool configuration
packages = []

[tool.ruff]
# Target Python version (Blender 5.0 uses Python 3.11)
target-version = "py311"

# Line length to match Black's default
line-length = 88

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".mypy_cache",
    ".ruff_cache",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable recommended rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "PL",  # Pylint
    "RUF", # Ruff-specific rules
]

ignore = [
    "E501",    # Line too long - handled by formatter
    "PLR0913", # Too many arguments - common in Blender APIs
    "PLR2004", # Magic value comparison - common in Blender
    "ARG002",  # Unused method argument - Blender API requires specific signatures
    "ARG003",  # Unused class method argument - Blender API requires specific signatures
    "RUF012",  # Mutable class attributes - Blender uses these for bl_options, etc.
    "PLC0415", # Import not at top level - needed for Blender addon reloading
    "TC002",   # Move import into type-checking block - fake-bpy types needed at runtime
]

[tool.ruff.lint.isort]
known-first-party = ["voxel_terrain"]
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # Allow unused imports in __init__.py

[tool.pyright]
# Python version (Blender 5.0 uses Python 3.11)
pythonVersion = "3.11"

# Type checking mode
typeCheckingMode = "standard"

# Report settings
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportDuplicateImport = true

# Blender modules are provided at runtime
reportMissingModuleSource = false

# Exclude from analysis
exclude = ["**/__pycache__", ".venv", ".git"]

# Include stubs from fake-bpy-module
stubPath = ""
